chore(core): Clean up hybrid impl
We can share `configurePlatform` with all implementations now, so make it private on the main type.

Consolidates other shared logic as well.
